nrf91: sync setup of west workspace with SDK
During west workspace setup we need to use newer version of NCS, as the
old version is not compatible with changes in SDK.

Update also how QCBOR is added as dependency to west.yml file (by
importing it from Golioth SDK manifest), so that its version does not
need to be specified explicitly.

Remove `west patch` step, as it was not required for long time and it
actually creates problems and confuses users.

Fixed: circuitdojo board name to match NCS 1.7.1 expectation