scripts: add mimxrt1060_evkb as custom board
Boards (and images) not available when scraping the Zephyr tree can now
be manually added. Place the yml and image in the
scripts/device-catalog/manually_added/ directory using a matching
directory structure to the zephyr tree. These will be copied into the
tree after git pull but before generating pages.

Signed-off-by: Mike Szczys <mike@golioth.io>