css: add Golioth brand colors and use them
This defines the Golioth color palette in custom CSS, and uses those define
colors to replace most of the color codes.

There should be no substantive appearance change as our initial site branding
was very close the official color palette. This does make the colors uniformed
and much easier to work with in the future.

There are a set of dark/er/est and light/er/est colors that should be reviewed
in a future commit.

Signed-off-by: Mike Szczys <mike@golioth.io>