scripts: fix closing slash detection
When using the redirect check, some sitemap links will have trailing slashes
while others will not. This happens when a directory URL loads an overview page.
This commit changes the detection method for the `{,/**}` filter. This is not
the most optimize approach; regex could have been used by we are splitting on
forward slashes so the pattern disrupted that split in previous versios.

Signed-off-by: Mike Szczys <mike@golioth.io>