scripts: correctly implement glob for redirects
Previously the `{,/**}` wildcard was not implementing the `**` part of the glob
in matching all folowing directories. This commit corrects that behavior.

The previous behavior broke the cascading nature of redirect parsing (redirects
not being applied to all child directories). This change fixes the issue, at the
expense of making accident sweeping redirects a bit harder to notice.

Printing of redirected links is on by default; scan carefully to ensure the
behavior you expect.

Signed-off-by: Mike Szczys <mike@golioth.io>