Zephry: add submodules init for Firmware SDK
When installing the Golioth Firmware SDK as the primary repository, you must use
git to update submodules or building will fail.

This is not necessary when the Golioth Firmware SDK is added as a submodule of
another project as the `submodules: true` entry in the manifest will perform
this action automatically.

Signed-off-by: Mike Szczys <mike@golioth.io>