device management: remove category files and overview frontmatter
- Use simple markdown title instead of frontmatter on the README overview pages
- Remove category files

The category files in this folder all incorrectly used the same sidebar
position. Label is the only other characteristic of these files that wasn't set
to the defaults. The titles of overview pages have been changes to match these
label values.

The frontmatter properties should only be used when necessary as defining a slug
become error prone when moving sections around. This commit moves to using folder
names for sidebar ordering, and markdown heading of the README.md pages that get
included as overview pages for sidebar entry names.

Signed-off-by: Mike Szczys <mike@golioth.io>