hardware-catalog: find images based on Zephyr's patterning
Use the same glob pattern approach from the Zephyr Doc generator to locate
images for the board catalog:

https://github.com/zephyrproject-rtos/zephyr/blob/999b19d6ceaee29902434bda9c160486a3eb75bc/doc/_scripts/gen_boards_catalog.py#L80-L100

This results in a small number of boards having an image selected that is not
the ideal image to use. This behavior matches what Zephyr is showing in Docs,
and a very large number of boards that didn't have a photo in our board catalog
now do.

Signed-off-by: Mike Szczys <mike@golioth.io>